Add files via upload
I describe my idea that house owner and house expect problem, that in our house we face issues in electricity, water leakage, woodwork and other household issues so that we need an expect so solve respective problems..I owner will difficulties to find the expect so we need an application where we have all the information about expects and their rating, price etc. So, I created this application.